rename to videotag_tsn_attention_lstm
