+++ hub_module/modules/video/classification/videotag_tsn_attention_lstm/resource/models/attention_lstm/lstm_attention.py
