+++ hub_module/modules/video/classification/videotag_tsn_attention_lstm/resource/predict.py
@@ -15,12 +15,12 @@
 import sys
 import paddle.fluid as fluid
 
-from VideoTag_TSN_AttentionLSTM.resource.utils.config_utils import *
-import VideoTag_TSN_AttentionLSTM.resource.models as models
-from VideoTag_TSN_AttentionLSTM.resource.reader import get_reader
-from VideoTag_TSN_AttentionLSTM.resource.metrics import get_metrics
-from VideoTag_TSN_AttentionLSTM.resource.utils.utility import check_cuda
-from VideoTag_TSN_AttentionLSTM.resource.utils.utility import check_version
+from videotag_tsn_attention_lstm.resource.utils.config_utils import *
+import videotag_tsn_attention_lstm.resource.models as models
+from videotag_tsn_attention_lstm.resource.reader import get_reader
+from videotag_tsn_attention_lstm.resource.metrics import get_metrics
+from videotag_tsn_attention_lstm.resource.utils.utility import check_cuda
+from videotag_tsn_attention_lstm.resource.utils.utility import check_version
 
 logging.root.handlers = []
 FORMAT = '[%(levelname)s: %(filename)s: %(lineno)4d]: %(message)s'
@@ -31,7 +31,7 @@ logger = logging.getLogger(__name__)
 def predict(args):
     """
     Video classification of 3k Chinese tags.
-    videoTag_TSN_AttentionLSTM (named as VideoTag_featureModel_predictModel)
+    videotag_tsn_attention_lstm (named as VideoTag_featureModel_predictModel)
     two stages in our model:
         1. extract feature from input video(mp4 format) using TSN model
         2. predict classification results from extracted feature  using AttentionLSTM model
