+++ hub_module/modules/video/classification/videotag_tsn_attention_lstm/resource/reader/kinetics_reader.py
